Refactor: 이메일 관련 기능 분리 및 리팩토링
이메일 관련 DTO 분리 및 생성
MemberDTO에서 이메일 관련 필드를 EmailDTO로 분리
이메일 인증에 필요한 데이터 모델을 EmailDTO로 새로 생성
EmailDTO에는 이메일 필드와 인증 코드 관련 필드를 포함
이메일 인증 처리를 위한 EmailService 구현
이메일 전송 및 인증 로직을 처리하는 EmailService 인터페이스 구현
이메일 전송, 인증 코드 생성, 인증 코드 검증 로직 포함
인증 코드 만료 및 이메일 중복 확인 기능 추가
이메일 전송 및 인증 로직을 포함한 EmailServiceImpl 추가
EmailServiceImpl 클래스에서 이메일 전송 및 인증 코드 검증 로직 구현
JavaMailSender를 사용한 이메일 전송 기능 추가
loadEmailProperties()와 configureJavaMailSender()를 사용하여 SMTP 설정 로딩
인증 코드 생성 및 이메일 전송 후 코드 저장 로직 추가
이메일 관련 엔드포인트를 위한 EmailController 생성
이메일 인증 관련 API 엔드포인트를 처리하는 EmailController 생성
/api/v1/member/email/code에서 인증 코드 전송 API 구현
/api/v1/member/email/verification에서 인증 코드 검증 API 구현
MemberService에서 EmailService로 이메일 관련 로직 이동
MemberService에서 이메일 관련 로직 제거
이메일 전송, 인증 코드 검증 등의 로직을 EmailService로 이전
EmailMapper 추가 및 이메일 중복 확인 기능을 EmailServiceImpl에서 처리
EmailServiceImpl에 로깅 추가
EmailServiceImpl의 주요 작업에 대해 로깅 추가
이메일 전송, 인증 코드 생성, 인증 코드 검증에 대한 로그 기록
오류 발생 시 log.error로 오류 메시지 출력
서비스 초기화 시 이메일 속성 로딩 수정
@PostConstruct 어노테이션을 사용하여 서비스 초기화 시 이메일 속성 로딩
loadEmailProperties()와 configureJavaMailSender()가 클래스 생성 후 자동으로 호출되도록 수정
