+++ src/main/java/fingertips/backend/member/controller/EmailController.java
@@ -1,7 +1,7 @@
 package fingertips.backend.member.controller;
 
 import fingertips.backend.exception.dto.JsonResponse;
-import fingertips.backend.member.dto.MemberDTO;
+import fingertips.backend.member.dto.EmailDTO;
 import fingertips.backend.member.service.EmailService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -21,24 +21,24 @@ public class EmailController {
 
     // 이메일 인증 코드 확인
     @PostMapping("/verification")
-    public ResponseEntity<JsonResponse<Boolean>> verifyEmail(@RequestBody MemberDTO memberDTO) {
-        boolean isVerified = emailService.verifyEmail(memberDTO.getEmail(), memberDTO.getInputCode());
+    public ResponseEntity<JsonResponse<Boolean>> verifyEmail(@RequestBody EmailDTO emailDTO) {
+        boolean isVerified = emailService.verifyEmail(emailDTO.getEmail(), emailDTO.getInputCode());
         return ResponseEntity.ok(JsonResponse.success(isVerified));
     }
 
     // 이메일 변경
     @PostMapping("/")
-    public ResponseEntity<JsonResponse<String>> requestEmailChange(@RequestBody MemberDTO memberDTO) {
+    public ResponseEntity<JsonResponse<String>> requestEmailChange(@RequestBody EmailDTO emailDTO) {
         String verificationCode = emailService.generateVerificationCode();
-        emailService.sendVerificationEmail(memberDTO.getNewEmail(), verificationCode);
+        emailService.sendVerificationEmail(emailDTO.getNewEmail(), verificationCode);
         return ResponseEntity.ok(JsonResponse.success("변경 요청이 성공적으로 전송되었습니다."));
     }
 
     // 인증 코드 전송
     @PostMapping("/code")
-    public ResponseEntity<JsonResponse<String>> sendVerificationCode(@RequestBody MemberDTO memberDTO) {
+    public ResponseEntity<JsonResponse<String>> sendVerificationCode(@RequestBody EmailDTO emailDTO) {
         String newVerificationCode = emailService.generateVerificationCode();
-        emailService.sendVerificationEmail(memberDTO.getEmail(), newVerificationCode);
+        emailService.sendVerificationEmail(emailDTO.getEmail(), newVerificationCode);
         return ResponseEntity.ok(JsonResponse.success("인증 코드가 전송되었습니다."));
     }
 }
