+++ src/main/java/fingertips/backend/member/dto/EmailDTO.java
@@ -1,0 +1,19 @@
+package fingertips.backend.member.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class EmailDTO {
+    private String email;
+    private String newEmail;
+    private String inputCode;
+    private LocalDateTime expirationTime;
+}
