+++ src/main/java/fingertips/backend/member/dto/MemberDTO.java
@@ -14,19 +14,16 @@ public class MemberDTO {
     private String memberId;
     private String password;
     private String memberName;
-    private String socialType;
-    private String socialId;
     private String birthday;
     private String gender;
     private String email;
-    private String newEmail;
-    private String inputCode;
     private String imageUrl;
     private String joinDate;
     private String region;
     private String job;
     private Integer testStatus;
     private Integer activeStatus;
+    private String socialType;
     private String refreshToken;
     private String withdrawDate;
     private String role;
