+++ src/main/java/fingertips/backend/member/mapper/EmailMapper.java
@@ -1,0 +1,6 @@
+package fingertips.backend.member.mapper;
+
+public interface EmailMapper {
+
+    int isEmailTaken(String email);
+}
