+++ src/main/java/fingertips/backend/member/service/MemberService.java
@@ -11,7 +11,6 @@ public interface MemberService {
     MemberDTO getMemberByMemberId(String memberId);
     void deleteMember(String username);
     void setRefreshToken(MemberDTO memberDTO);
-    boolean isEmailTaken(String email);
     boolean existsMemberId(String memberId);
     void clearRefreshToken(String memberId);
     String findByNameAndEmail(String memberName, String email);
