+++ src/main/resources/mapper/EmailMapper.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+
+<mapper namespace="fingertips.backend.member.mapper.EmailMapper">
+    <select id="isEmailTaken" parameterType="String" resultType="Integer">
+        SELECT COUNT(*)
+        FROM member
+        WHERE email = #{email}
+    </select>
+</mapper>
