+++ src/main/resources/mapper/MemberMapper.xml
@@ -34,12 +34,6 @@
           AND member_name = #{memberName}
     </select>
 
-    <select id="isEmailTaken" parameterType="String" resultType="Integer">
-        SELECT COUNT(*)
-        FROM member
-        WHERE email = #{email}
-    </select>
-
     <select id="existsMemberId" parameterType="String" resultType="Integer">
         SELECT COUNT(*)
         FROM member
